tg_bot/requirements.txt
aiogram==3.13.0
python-dotenv==1.0.0
httpx==0.27.2
requests==2.32.3